requirements.txt
resim-open-core